# Avoid using Lazy<> for drawable scheduler to reduce alloc overhead