Due to the reference to `this.Clock`, the delegate was being constructed per drawable. Not a huge overhead but I'd rather not see it while profiling (as such large numbers of `Func` usually point to an issue).

![image](https://user-images.githubusercontent.com/191335/95650097-3d473a00-0b1c-11eb-88ae-7f38594225ca.png)

Benchmarked rebased on #3927:

before:
|              Method |     Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|--------:|--------:|------:|------:|------:|----------:|
|   TestEmptyGameLoop | 994.5 us | 2.52 us | 2.36 us |     - |     - |     - |      41 B |
|  TestAllocContainer | 993.4 us | 2.03 us | 1.90 us |     - |     - |     - |    1996 B |
| TestLoadedContainer | 993.0 us | 1.57 us | 1.47 us |     - |     - |     - |    2404 B |
|     TestAllocSprite | 992.3 us | 7.51 us | 7.02 us |     - |     - |     - |    1601 B |
|    TestLoadedSprite | 993.8 us | 1.66 us | 1.56 us |     - |     - |     - |    1853 B |

after:
|              Method |     Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|--------:|--------:|------:|------:|------:|----------:|
|   TestEmptyGameLoop | 992.4 us | 1.41 us | 1.25 us |     - |     - |     - |      41 B |
|  TestAllocContainer | 992.4 us | 2.38 us | 2.23 us |     - |     - |     - |    1706 B |
| TestLoadedContainer | 992.6 us | 1.72 us | 1.61 us |     - |     - |     - |    2132 B |
|     TestAllocSprite | 992.9 us | 2.42 us | 2.26 us |     - |     - |     - |    1484 B |
|    TestLoadedSprite | 992.4 us | 3.00 us | 2.80 us |     - |     - |     - |    1717 B |