@peppy 2020-10-11
Added benchmark coverage to OP.